using input system now

+++ Assets/Scripts/Queens/Models/CardModel.cs
@@ -12,14 +12,14 @@ namespace Queens.Models
         public string dialog{ get; set; }
         public string yes_answer{ get; set; }
         public string no_answer{ get; set; }
-        public float? yes_flow{ get; set; }
-        public float? yes_popularity{ get; set; }
-        public float? yes_money{ get; set; }
-        public float? yes_health{ get; set; }
-        public float? no_flow{ get; set; }
-        public float? no_popularity{ get; set; }
-        public float? no_money{ get; set; }
-        public float? no_health{ get; set; }
+        public int? yes_flow{ get; set; }
+        public int? yes_popularity{ get; set; }
+        public int? yes_money{ get; set; }
+        public int? yes_health{ get; set; }
+        public int? no_flow{ get; set; }
+        public int? no_popularity{ get; set; }
+        public int? no_money{ get; set; }
+        public int? no_health{ get; set; }
         public int? level_lock{ get; set; }
     }
 }
